+++ clients/admin-ui/src/features/configure-consent/AddVendor.tsx
@@ -21,7 +21,11 @@ import {
   selectAllDictEntries,
   useGetAllDictionaryEntriesQuery,
 } from "~/features/plus/plus.slice";
-import { useCreateSystemMutation } from "~/features/system";
+import {
+  useCreateSystemMutation,
+  useUpdateSystemMutation,
+} from "~/features/system";
+import { System } from "~/types/api";
 
 import { getErrorMessage, isErrorResult } from "../common/helpers";
 import { errorToastParams, successToastParams } from "../common/toast";
@@ -29,7 +33,7 @@ import AddModal from "./AddModal";
 import { EMPTY_DECLARATION, FormValues } from "./constants";
 import DataUsesForm from "./DataUsesForm";
 
-const initialValues: FormValues = {
+const defaultInitialValues: FormValues = {
   name: "",
   vendor_id: undefined,
   privacy_declarations: [EMPTY_DECLARATION],
@@ -43,8 +47,18 @@ const DictionaryValidationSchema = Yup.object().shape({
   vendor_id: Yup.string().required().label("Vendor"),
 });
 
-const AddVendor = () => {
-  const modal = useDisclosure();
+const AddVendor = ({
+  passedInSystem,
+  onCloseModal,
+}: {
+  passedInSystem?: System;
+  onCloseModal?: () => void;
+}) => {
+  const defaultModal = useDisclosure();
+  const modal = {
+    ...defaultModal,
+    isOpen: passedInSystem ? true : defaultModal.isOpen,
+  };
   const toast = useToast();
 
   // Subscribe and get dictionary values
@@ -56,12 +70,31 @@ const AddVendor = () => {
   const dictionaryOptions = useAppSelector(selectAllDictEntries);
 
   const [createSystemMutationTrigger] = useCreateSystemMutation();
+  const [updateSystemMutationTrigger] = useUpdateSystemMutation();
   const [isShowingSuggestions, setIsShowingSuggestions] = useState(false);
 
   const handleCloseModal = () => {
     modal.onClose();
+    if (onCloseModal) {
+      onCloseModal();
+    }
     setIsShowingSuggestions(false);
   };
+
+  const initialValues = passedInSystem
+    ? {
+        name: passedInSystem.name ?? "",
+        vendor_id: passedInSystem.vendor_id,
+        privacy_declarations: passedInSystem.privacy_declarations.map(
+          (dec) => ({
+            ...dec,
+            name: dec.name ?? "",
+            cookies: dec.cookies ?? [],
+            cookieNames: dec.cookies ? dec.cookies.map((c) => c.name) : [],
+          })
+        ),
+      }
+    : defaultInitialValues;
 
   const handleSubmit = async (
     values: FormValues,
@@ -70,16 +103,14 @@ const AddVendor = () => {
     const transformedDeclarations = values.privacy_declarations
       .filter((dec) => dec.data_use !== EMPTY_DECLARATION.data_use)
       .map((dec) => {
-        // `cookieNames` is only a temporary field to help with the form
-        // At this point, we should see if we need to transform `cookieNames`,
-        // or if we can work with `cookies` directly.
-        let transformedCookies = dec.cookies;
-        if (dec.cookies.length === 0) {
-          transformedCookies = dec.cookieNames.map((name) => ({
-            name,
-            path: "/",
-          }));
-        }
+        // if a cookie from the form already exists on the declaration with full
+        // information from the dictionary, use that; otherwise, make the cookie
+        // name from the form into a new cookie
+        const transformedCookies = dec.cookieNames.map((name) => {
+          const existingCookie = dec.cookies.find((c) => c.name === name);
+          return existingCookie ?? { name, path: "/" };
+        });
+
         const { cookieNames, ...rest } = dec;
         return { ...rest, cookies: transformedCookies };
       });
@@ -88,27 +119,35 @@ const AddVendor = () => {
     // so now we need to clear out vendor_id if it's not a system in the dictionary
     const vendor = dictionaryOptions.find((o) => o.value === values.vendor_id);
     const vendorId = vendor ? vendor.value : undefined;
-    let { name } = values;
+    let { name: newName } = values;
     if (vendor) {
-      name = vendor.label;
+      newName = vendor.label;
     } else if (values.vendor_id) {
       // This is the case where the user created their own vendor name
-      name = values.vendor_id;
+      newName = values.vendor_id;
     }
 
     const payload = {
       vendor_id: vendorId,
-      name,
-      fides_key: formatKey(name),
-      system_type: "",
+      name: passedInSystem ? passedInSystem.name : newName,
+      fides_key: passedInSystem ? passedInSystem.fides_key : formatKey(newName),
+      system_type: passedInSystem ? passedInSystem.system_type : "",
       privacy_declarations: transformedDeclarations,
     };
-    const result = await createSystemMutationTrigger(payload);
+
+    const result = passedInSystem
+      ? await updateSystemMutationTrigger(payload)
+      : await createSystemMutationTrigger(payload);
+
     if (isErrorResult(result)) {
       toast(errorToastParams(getErrorMessage(result.error)));
       return;
     }
-    toast(successToastParams("Vendor successfully added!"));
+    toast(
+      successToastParams(
+        `Vendor successfully ${passedInSystem ? "updated" : "created"}!`
+      )
+    );
     helpers.resetForm();
     handleCloseModal();
   };
@@ -119,7 +158,12 @@ const AddVendor = () => {
 
   return (
     <>
-      <Button onClick={modal.onOpen} data-testid="add-vendor-btn">
+      <Button
+        onClick={modal.onOpen}
+        data-testid="add-vendor-btn"
+        size="sm"
+        colorScheme="primary"
+      >
         Add vendor
       </Button>
       <Formik
@@ -139,15 +183,15 @@ const AddVendor = () => {
             <AddModal
               isOpen={modal.isOpen}
               onClose={modal.onClose}
-              title="Add a vendor"
+              title={passedInSystem ? "Edit vendor" : "Add a vendor"}
               onSuggestionClick={() => {
                 setIsShowingSuggestions(true);
               }}
               suggestionsState={suggestionsState}
             >
-              <Box data-testid="add-vendor-modal-content">
+              <Box data-testid="add-vendor-modal-content" my={4}>
                 <Form>
-                  <VStack alignItems="start">
+                  <VStack alignItems="start" spacing={6}>
                     {hasDictionary ? (
                       <CustomCreatableSelect
                         id="vendor"
@@ -159,9 +203,12 @@ const AddVendor = () => {
                         tooltip="Select the vendor that matches the system"
                         isCustomOption
                         variant="stacked"
+                        isDisabled={!!passedInSystem}
                         isRequired
                       />
-                    ) : (
+                    ) : null}
+                    {(passedInSystem && !passedInSystem.vendor_id) ||
+                    !hasDictionary ? (
                       <CustomTextInput
                         id="name"
                         name="name"
@@ -169,8 +216,9 @@ const AddVendor = () => {
                         label="Vendor name"
                         tooltip="Give the system a unique, and relevant name for reporting purposes. e.g. “Email Data Warehouse”"
                         variant="stacked"
+                        disabled={!!passedInSystem}
                       />
-                    )}
+                    ) : null}
                     <DataUsesForm showSuggestions={isShowingSuggestions} />
                     <ButtonGroup
                       size="sm"